+++ Data/ByteString.hs
@@ -159,6 +159,7 @@ module Data.ByteString (
         elemIndexEnd,           -- :: Word8 -> ByteString -> Maybe Int
         findIndex,              -- :: (Word8 -> Bool) -> ByteString -> Maybe Int
         findIndices,            -- :: (Word8 -> Bool) -> ByteString -> [Int]
+        findIndexEnd,           -- :: (Word8 -> Bool) -> ByteString -> Maybe Int
         count,                  -- :: Word8 -> ByteString -> Int
 
         -- * Zipping and unzipping ByteStrings
@@ -1112,14 +1113,7 @@ elemIndex c (PS x s l) = accursedUnutterablePerformIO $ withForeignPtr x $ \p ->
 -- > (-) (length xs - 1) `fmap` elemIndex c (reverse xs)
 --
 elemIndexEnd :: Word8 -> ByteString -> Maybe Int
-elemIndexEnd ch (PS x s l) = accursedUnutterablePerformIO $ withForeignPtr x $ \p ->
-    go (p `plusPtr` s) (l-1)
-  where
-    go !p !i | i < 0     = return Nothing
-             | otherwise = do ch' <- peekByteOff p i
-                              if ch == ch'
-                                  then return $ Just i
-                                  else go p (i-1)
+elemIndexEnd = findIndexEnd . (==)
 {-# INLINE elemIndexEnd #-}
 
 -- | /O(n)/ The 'elemIndices' function extends 'elemIndex', by returning
@@ -1148,7 +1142,7 @@ count w (PS x s m) = accursedUnutterablePerformIO $ withForeignPtr x $ \p ->
     fmap fromIntegral $ c_count (p `plusPtr` s) (fromIntegral m) w
 {-# INLINE count #-}
 
--- | The 'findIndex' function takes a predicate and a 'ByteString' and
+-- | /O(n)/ The 'findIndex' function takes a predicate and a 'ByteString' and
 -- returns the index of the first element in the ByteString
 -- satisfying the predicate.
 findIndex :: (Word8 -> Bool) -> ByteString -> Maybe Int
@@ -1161,7 +1155,20 @@ findIndex k (PS x s l) = accursedUnutterablePerformIO $ withForeignPtr x $ \f ->
                                   else go (ptr `plusPtr` 1) (n+1)
 {-# INLINE findIndex #-}
 
--- | The 'findIndices' function extends 'findIndex', by returning the
+-- | /O(n)/ The 'findIndexEnd' function takes a predicate and a 'ByteString' and
+-- returns the index of the last element in the ByteString
+-- satisfying the predicate.
+findIndexEnd :: (Word8 -> Bool) -> ByteString -> Maybe Int
+findIndexEnd k (PS x s l) = accursedUnutterablePerformIO $ withForeignPtr x $ \ f -> go (f `plusPtr` s) (l-1)
+  where
+    go !ptr !n | n < 0     = return Nothing
+               | otherwise = do w <- peekByteOff ptr n
+                                if k w
+                                  then return (Just n)
+                                  else go ptr (n-1)
+{-# INLINE findIndexEnd #-}
+
+-- | /O(n)/ The 'findIndices' function extends 'findIndex', by returning the
 -- indices of all elements satisfying the predicate, in ascending order.
 findIndices :: (Word8 -> Bool) -> ByteString -> [Int]
 findIndices p ps = loop 0 ps